requirements/base.txt
django

django-environ
whitenoise

psycopg2

pytz

django-redis
redis

djangorestframework
django-rest-swagger
djoser
